Added Royd's updated files from the branch, changed how main menu and edit entry views are launched
Main menu now accepts passed-in file names to access list of entry titles. This allows information to be transferred seamlessly between scenes when going back and forth. Also added parameter to allow file name of complete list of entry names to be passed to edit entry scene so this can be passed back to the main menu when returning. This should make it easier to save updated entry names when setting up save and create functions.