Added save capabilities
EntryController now allows the user to save an entry as long as the entry title is at least one character long. If the entry title matches an existing entry title, a pop-up alert confirms the user wishes to overwrite the prior save. MenuController now allows user to create a new entry by switching to the Edit Entry view with a blank form. Removed redundant menu bar components that are taken care of by buttons on the GUI.